+++ src/instructions/traps/puts.rs
@@ -3,14 +3,15 @@ use std::io::{stdout, Write};
 use crate::{memory::Memory, registers::Register};
 
 pub fn puts(reg: &mut [u16], mem: &mut Memory) {
-    let mut index = reg[Register::RR0 as usize] as usize;
-    let mut c = mem.mem[index];
+    let mut index = reg[Register::RR0 as usize];
+
+    let mut c = mem.mem[index as usize];
 
     while c != 0x0000 {
         print!("{}", (c as u8) as char);
         index += 1;
 
-        c = mem.mem[index];
+        c = mem.mem[index as usize];
     }
     stdout().flush().expect("Flushed");
 }
